se migro a mysql

+++ database/migrations/2016_04_08_023641_create_vehiculo_table.php
@@ -27,7 +27,7 @@ class CreateVehiculoTable extends Migration
 
             //$table->bigInteger('conductor_cedula')->unsigned();
             //$table->foreign('conductor_cedula')->references('cedula')->on('conductor')->onDelete('cascade');
-            $table->bigInteger('conductor_id')->unsigned();
+            $table->integer('conductor_id')->unsigned();
             $table->foreign('conductor_id')->references('id')->on('conductor')->onDelete('cascade');
 
             $table->timestamps();
